fix: restructured LeaderBoard component
