--- packages/client/src/components/back-button/index.ts
@@ -1,1 +1,0 @@
-export * from './BackButton'
